+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Truncate file names in the middle, not at the end; important information are more often at the end
 * Remove email address from user profile
 * Mark external links with " ↗" to make them clear
+* Make QR code larger on "Add Second Device" screen
 * Fix: Show dialog if pasted QR codes are invalid
 * Update to core 2.36.0